field: fix heap corruption from FieldImageConvert24BitTo15Bit u_long overflow
After the OT fix the field reached func_800A5774's HeapAlloc, which crashed
in HeapConsolidate walking a free list with a garbage pNext (0x2a74). A gdb
hardware watchpoint traced the bad write to FieldImageConvert24BitTo15Bit
(misc5.c) overflowing its pImage15Bit buffer into adjacent heap headers.

The convert loop runs 0x1C00 (7168) iterations doing
  *g_Field15BitImageData = ...; g_Field15BitImageData += 1;
g_Field15BitImageData is u_long*. On PSX u_long is 4 bytes, so 7168*4 = 0x7000
exactly fills the HeapAlloc(0x7000) buffer. The 64-bit port widens u_long to 8
bytes, so it writes 7168*8 = 0xE000 -- 0x7000 bytes past the buffer, clobbering
the next blocks' pNext fields. The 24-bit read side (g_Field24BitImageData) was
broken identically.

Fix: type these image buffers as u32 instead of u_long so the deref/stride stays
4 bytes. u32 == u_long == 4 bytes on the MIPS target, so codegen is byte-identical
and matching is preserved; on the 64-bit host u32 is correct. Same class of bug as
the OT_TAG stride fix -- a u_long 8-byte widening port-ism.

Verified: field clears the convert pass and enters the FieldMain main loop
([FieldMain] entering main loop); new frontier is FieldLoad (g_FieldActors NULL,
already scoped). KernelMenu still reaches the sustained loop (no regression).

+++ src/field/main/misc5.c
@@ -172,12 +172,17 @@ INCLUDE_ASM("asm/field/nonmatchings/main/misc5", func_800A73E8);
 INCLUDE_ASM("asm/field/nonmatchings/main/misc5", func_800A74F8);
 
 extern int g_FieldPixelIndex;
-extern u_long g_FieldCurPixel;
-extern u_long* g_Field24BitImageData;
-extern u_long* g_Field15BitImageData;
+/* These image-convert buffers are 32-bit words on PSX (u_long == u32 == 4 bytes
+ * there). The 64-bit port widens u_long to 8 bytes, which would double the deref
+ * width and pointer stride below -- overflowing pImage15Bit (0x7000 bytes, written
+ * 0x1C00 times) into adjacent heap blocks. Use u32 so the stride stays 4 bytes;
+ * byte-identical to u_long on the MIPS target, so matching is preserved. */
+extern u32 g_FieldCurPixel;
+extern u32* g_Field24BitImageData;
+extern u32* g_Field15BitImageData;
 
 u_int FieldImageConvert24BitColorTo15Bit(void) {
-    u_long nPixel;
+    u32 nPixel;
     u_int nLSB;
 
     // Are we done reading RGB channels?
@@ -205,9 +210,9 @@ void FieldImageConvert24BitTo15Bit(void) {
     RECT rect;
     int i;
     int j;
-    u_long* pImage24Bit;
-    u_long* pImage15Bit;
-    u_long n15BitPixels;
+    u32* pImage24Bit;
+    u32* pImage15Bit;
+    u32 n15BitPixels;
 
     pImage24Bit = HeapAlloc(0xA800, 0x0);
     pImage15Bit = HeapAlloc(0x7000, 0x0);
